+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+1.14.0
+-----
+* Support standard HTML colors by name.
+* Add support for reverse and start attributes for ordered lists.
+
 1.13.0
 -----
 * Fix a bug when copying and pasting from other apps when running dark mode.